Update .net version to 6.0, some warning fixes, and directories
