+++ AspNetCore.Yandex.ObjectStorage/AspNetCore.Yandex.ObjectStorage.csproj
@@ -13,12 +13,12 @@
     <PackageRequireLicenseAcceptance>false</PackageRequireLicenseAcceptance>
     <SignAssembly>false</SignAssembly>
     <PackageIconUrl>https://yastatic.net/q/cloud-www/build/ru/assets/img/storage.15691e8c.png</PackageIconUrl>
-    <AssemblyVersion>0.1.7.6</AssemblyVersion>
-    <FileVersion>0.1.7.6</FileVersion>
+    <AssemblyVersion>0.1.8.0</AssemblyVersion>
+    <FileVersion>0.1.8.0</FileVersion>
     <PackageLicenseFile>LICENSE.md</PackageLicenseFile>
     <EnableNETAnalyzers>true</EnableNETAnalyzers>
-    <PackageVersion>0.1.7.6</PackageVersion>
-    <TargetFrameworks>net5.0;netcoreapp3.1</TargetFrameworks>
+    <PackageVersion>0.1.8.0</PackageVersion>
+    <TargetFrameworks>net5.0;net6.0;netcoreapp3.1</TargetFrameworks>
   </PropertyGroup>
 
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
@@ -27,9 +27,9 @@
 
   <ItemGroup>
     <PackageReference Include="Microsoft.AspNetCore.WebUtilities" Version="2.2.0" />
-    <PackageReference Include="Microsoft.Extensions.Configuration" Version="5.0.0" />
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="5.0.0" />
-    <PackageReference Include="Microsoft.Extensions.Options" Version="5.0.0" />
+    <PackageReference Include="Microsoft.Extensions.Configuration" Version="6.0.0" />
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="6.0.0" />
+    <PackageReference Include="Microsoft.Extensions.Options" Version="6.0.0" />
   </ItemGroup>
 
   <ItemGroup>
