+++ AspNetCore.Yandex.ObjectStorage/Helpers/PathHelper.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Text.RegularExpressions;
 
-namespace AspNetCore.Yandex.ObjectStorage
+namespace AspNetCore.Yandex.ObjectStorage.Helpers
 {
 	public static class PathHelper
 	{
@@ -18,7 +18,7 @@ namespace AspNetCore.Yandex.ObjectStorage
 		public static string RemoveBucket(this string path, string bucket)
 		{
 			var regex = new Regex(Regex.Escape($"{bucket}/"));
-			return regex.Replace(path, $"", 1);
+			return regex.Replace(path, "", 1);
 		}
 
 		public static int IndexOfFile(this string path)
@@ -26,7 +26,6 @@ namespace AspNetCore.Yandex.ObjectStorage
 			if (string.IsNullOrWhiteSpace(path)) throw new ArgumentNullException(nameof(path));
 
 			var formatIndex = path.LastIndexOf('.');
-
 			if (formatIndex > 0) // Then file exist in path
 			{
 				return path.Contains('/') ? path.LastIndexOf('/') : 0;
